✨ isTreePath
